add metrics
